[Silver V] Title: 타일 장식물, Time: 128 ms, Memory: 14288 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.java
@@ -6,17 +6,17 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
-        int answer = 0;
+        long answer = 0;
         if(n == 1) answer = 4;
         else if(n == 2) answer = 6;
         else {
-            int[] dp = new int[n + 1];
+            long[] dp = new long[n + 1];
             dp[1] = 1;
             dp[2] = 2;
             for (int i = 3; i <= n; i++) {
                 dp[i] = dp[i - 1] + dp[i - 2];
             }
-            answer = 2 * (dp[n-1] + dp[n]);
+            answer = 2L * (dp[n-1] + dp[n]);
         }
         System.out.println(answer);
     }
